fix: remove HostServices (only working with Oracle JRE)
This makes java.awt.Desktop the default (and only) option to attempt opening links in the system browser.
